Add unit tests for subset of kernels (#1384)
* pr for debugging kernel driver issues

* Commit flake files

* current kernel tests

* softmax, flat, transpose kernel tests

* clang formatting kernel tests

* reverse, split, full dropout kernels

* rest of kernel-tests

* minor cleannup

* Restore .proj.toml

* Delete misadded directory

* merge fix

* more merge fixes

* resolved merge conflicts with repo-refactor

* code review changes

* allocator updates

* allocation util updates

* test clean up and review fixes

* fixed forward backward pass consistencies, added filler tests for all tests, other review changes

* unnested test subcases and more review changes

* added managed_stream and handle classes, other minor clean up

* fix accessor and corresponding shape clarity, other clean up

* merge error fixes

* managed handle and stream fixes, removed datatype dispatch from cuda_helper, other clean up

* managed handle and stream updates

* fixed deallocator

---------

+++ lib/local-execution/src/ops/attention.cc
@@ -84,10 +84,10 @@ static DeviceSpecific<DeviceStates>
     init_task_impl(TaskArgumentAccessor const &acc) {
   auto const &attrs = acc.get_argument<MultiHeadAttentionAttrs>(ATTRS);
   Allocator allocator = acc.get_allocator();
-  int qProjSize = acc.get_argument<int>(QPROJSIZE);
-  int kProjSize = acc.get_argument<int>(KPROJSIZE);
-  int vProjSize = acc.get_argument<int>(VPROJSIZE);
-  int oProjSize = acc.get_argument<int>(OPROJSIZE);
+  size_t qProjSize = acc.get_argument<size_t>(QPROJSIZE);
+  size_t kProjSize = acc.get_argument<size_t>(KPROJSIZE);
+  size_t vProjSize = acc.get_argument<size_t>(VPROJSIZE);
+  size_t oProjSize = acc.get_argument<size_t>(OPROJSIZE);
   PerDeviceFFHandle handle = acc.get_argument<PerDeviceFFHandle>(HANDLE);
   ParallelTensorShape query_parallel_tensor_shape =
       acc.get_argument<ParallelTensorShape>(QUERY_PARALLEL_TENSOR_SHAPE);
